requirements.txt
asgiref==3.4.1
cffi==1.14.6
cryptography==35.0.0
cycler==0.10.0
Django==3.2.7
django-cors-headers==3.9.0
django-picklefield==3.0.1
django-rest-knox==4.1.0
djangorestframework==3.12.4
joblib==1.0.1
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.2
pandas==1.3.3
Pillow==8.3.2
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
scikit-learn==1.0
scipy==1.7.1
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.0.0
